@@ -19,9 +19,10 @@ disallow_untyped_defs = "True"
 "Release Notes" = "https://github.com/langchain-ai/langchain/releases?q=tag%3A%22langchain-groq%3D%3D0%22&expanded=true"
 
 [tool.poetry.dependencies]
-python = ">=3.8.1,<4.0"
+python = ">=3.9,<4.0"
 langchain-core = "^0.2.26"
 groq = ">=0.4.1,<1"
+pydantic = ">=2,<3"
 
 [tool.ruff.lint]
 select = [ "E", "F", "I", "W",]
