@@ -25,7 +25,7 @@ ignore_missing_imports = "True"
 "Release Notes" = "https://github.com/langchain-ai/langchain/releases?q=tag%3A%22langchain-milvus%3D%3D0%22&expanded=true"
 
 [tool.poetry.dependencies]
-python = ">=3.8.1,<4.0"
+python = ">=3.9,<4.0"
 langchain-core = "^0.2.38"
 pymilvus = "^2.4.3"
 [[tool.poetry.dependencies.scipy]]
