@@ -12,7 +12,7 @@ license = "MIT"
 "Release Notes" = "https://github.com/langchain-ai/langchain/releases?q=tag%3A%22langchain-nomic%3D%3D0%22&expanded=true"
 
 [tool.poetry.dependencies]
-python = ">=3.8.1,<4.0"
+python = ">=3.9,<4.0"
 langchain-core = ">=0.1.46,<0.3"
 nomic = "^3.0.29"
 pillow = "^10.3.0"
