@@ -66,17 +66,19 @@ from langchain_core.output_parsers.openai_tools import (
     parse_tool_call,
 )
 from langchain_core.outputs import ChatGeneration, ChatGenerationChunk, ChatResult
-from langchain_core.pydantic_v1 import (
-    BaseModel,
-    Field,
-    SecretStr,
-    root_validator,
-)
 from langchain_core.runnables import Runnable, RunnableMap, RunnablePassthrough
 from langchain_core.tools import BaseTool
 from langchain_core.utils import secret_from_env
 from langchain_core.utils.function_calling import convert_to_openai_tool
 from langchain_core.utils.pydantic import is_basemodel_subclass
+from pydantic import (
+    BaseModel,
+    ConfigDict,
+    Field,
+    SecretStr,
+    model_validator,
+)
+from typing_extensions import Self
 
 logger = logging.getLogger(__name__)
 
@@ -379,11 +381,10 @@ class ChatMistralAI(BaseChatModel):
     safe_mode: bool = False
     streaming: bool = False
 
-    class Config:
-        """Configuration for this pydantic object."""
-
-        allow_population_by_field_name = True
-        arbitrary_types_allowed = True
+    model_config = ConfigDict(
+        populate_by_name=True,
+        arbitrary_types_allowed=True,
+    )
 
     @property
     def _default_params(self) -> Dict[str, Any]:
@@ -469,47 +470,50 @@ class ChatMistralAI(BaseChatModel):
         combined = {"token_usage": overall_token_usage, "model_name": self.model}
         return combined
 
-    @root_validator(pre=False, skip_on_failure=True)
-    def validate_environment(cls, values: Dict) -> Dict:
+    @model_validator(mode="after")
+    def validate_environment(self) -> Self:
         """Validate api key, python package exists, temperature, and top_p."""
-        api_key_str = values["mistral_api_key"].get_secret_value()
+        if isinstance(self.mistral_api_key, SecretStr):
+            api_key_str: Optional[str] = self.mistral_api_key.get_secret_value()
+        else:
+            api_key_str = self.mistral_api_key
 
         # todo: handle retries
         base_url_str = (
-            values.get("endpoint")
+            self.endpoint
             or os.environ.get("MISTRAL_BASE_URL")
             or "https://api.mistral.ai/v1"
         )
-        values["endpoint"] = base_url_str
-        if not values.get("client"):
-            values["client"] = httpx.Client(
+        self.endpoint = base_url_str
+        if not self.client:
+            self.client = httpx.Client(
                 base_url=base_url_str,
                 headers={
                     "Content-Type": "application/json",
                     "Accept": "application/json",
                     "Authorization": f"Bearer {api_key_str}",
                 },
-                timeout=values["timeout"],
+                timeout=self.timeout,
             )
         # todo: handle retries and max_concurrency
-        if not values.get("async_client"):
-            values["async_client"] = httpx.AsyncClient(
+        if not self.async_client:
+            self.async_client = httpx.AsyncClient(
                 base_url=base_url_str,
                 headers={
                     "Content-Type": "application/json",
                     "Accept": "application/json",
                     "Authorization": f"Bearer {api_key_str}",
                 },
-                timeout=values["timeout"],
+                timeout=self.timeout,
             )
 
-        if values["temperature"] is not None and not 0 <= values["temperature"] <= 1:
+        if self.temperature is not None and not 0 <= self.temperature <= 1:
             raise ValueError("temperature must be in the range [0.0, 1.0]")
 
-        if values["top_p"] is not None and not 0 <= values["top_p"] <= 1:
+        if self.top_p is not None and not 0 <= self.top_p <= 1:
             raise ValueError("top_p must be in the range [0.0, 1.0]")
 
-        return values
+        return self
 
     def _generate(
         self,
@@ -728,7 +732,7 @@ class ChatMistralAI(BaseChatModel):
                 from typing import Optional
 
                 from langchain_mistralai import ChatMistralAI
-                from langchain_core.pydantic_v1 import BaseModel, Field
+                from pydantic import BaseModel, Field
 
 
                 class AnswerWithJustification(BaseModel):
@@ -759,7 +763,7 @@ class ChatMistralAI(BaseChatModel):
             .. code-block:: python
 
                 from langchain_mistralai import ChatMistralAI
-                from langchain_core.pydantic_v1 import BaseModel
+                from pydantic import BaseModel
 
 
                 class AnswerWithJustification(BaseModel):
@@ -846,7 +850,7 @@ class ChatMistralAI(BaseChatModel):
             .. code-block::
 
                 from langchain_mistralai import ChatMistralAI
-                from langchain_core.pydantic_v1 import BaseModel
+                from pydantic import BaseModel
 
                 class AnswerWithJustification(BaseModel):
                     answer: str
